renamed to nucos, and refactored setup.py
