+++ nucos/__init__.py
@@ -10,9 +10,7 @@ All of unit_conversion is imported here for convenience
 
 """
 
-from __future__ import division, absolute_import, unicode_literals
-
-__version__ = "2.12.0"
+__version__ = "3.0.0"
 
 from .unit_conversion import *
