+++ setup.cfg
@@ -1,0 +1,26 @@
+[metadata]
+name = pynucos
+version = attr: nucos.__version__
+author = Christopher H. Barker
+author_email = Chris.Barker@noaa.gov
+description = Physical Unit conversion utilities -- units useful for oil and chemical spill response'
+long_description = file: README.rst
+long_description_content_type = text/x-rst
+url = https://github.com/NOAA-ORR-ERD/PyNUCOS
+project_urls =
+    Bug Tracker = https://github.com/NOAA-ORR-ERD/PyNUCOS/issues
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: MIT License
+    Operating System :: OS Independent
+    Development Status :: 5 - Production/Stable
+    Intended Audience :: Developers
+    Intended Audience :: Science/Research
+    License :: CC0 1.0 Universal (CC0 1.0) Public Domain Dedication
+    Programming Language :: Python :: 3 :: Only
+    Topic :: Scientific/Engineering
+
+[options]
+packages = find:
+python_requires = >=3.7
+zip_safe = False
